requirement.txt
absl-py==1.4.0
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dm-tree==0.1.9
flatbuffers==25.2.10
fonttools==4.59.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.73.1
gTTS==2.5.4
h5py==3.14.0
idna==3.10
importlib-metadata==6.11.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==2.12.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.1
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
narwhals==1.47.0
numpy==1.23.5
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.16.0
packaging==23.2
pandas==2.3.1
Pillow==9.5.0
playsound==1.3.0
protobuf==4.25.8
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
pywin32==311
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.2.2
scipy==1.15.3
seaborn==0.12.2
six==1.17.0
smmap==5.0.2
SpeechRecognition==3.14.3
streamlit==1.23.1
tenacity==8.5.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-model-optimization==0.8.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==4.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.14.1
zipp==3.23.0